add CSV file

+++ database/migrations/2020_11_05_082109_create_persons_table.php
@@ -18,7 +18,7 @@ class CreatePersonsTable extends Migration
             $table->string('name');
             $table->Integer('age');
             $table->Integer('phone');
-            $table->Integer('monthly_fee');
+            $table->Double('monthly_fee');
             $table->string('month');
         });
     }
